Move integration tests of Spring Boot app example under "test" directory
