+++ detekt.yml
@@ -7,12 +7,6 @@ comments:
     active: false
   CommentOverPrivateFunction:
     active: false
-  UndocumentedPublicClass:
-    excludes: "**/test/**,**\\test\\**,**/*Test.kt,**/*Spec.kt,**/*IT.kt,**/*TestConfiguration.kt"
-  UndocumentedPublicFunction:
-    excludes: "**/test/**,**\\test\\**,**/*Test.kt,**/*Spec.kt,**/*IT.kt,**/*TestConfiguration.kt"
-  UndocumentedPublicProperty:
-    excludes: "**/test/**,**\\test\\**,**/*Test.kt,**/*Spec.kt,**/*IT.kt,**/*TestConfiguration.kt"
 
 complexity:
   LongParameterList:
@@ -28,7 +22,6 @@ complexity:
 naming:
   FunctionMaxLength:
     maximumFunctionNameLength: 35
-    excludes: "**/test/**,**\\test\\**,**/*Test.kt,**/*Spec.kt,**/*IT.kt,**/*TestConfiguration.kt"
 
 style:
   MagicNumber:
