+++ examples/spring/src/test/kotlin/com/expediagroup/graphql/examples/IntegrationTestConstants.kt
@@ -18,12 +18,9 @@ package com.expediagroup.graphql.examples
 
 import org.springframework.http.MediaType
 
-object Constants {
-
-    const val DATA_JSON_PATH = "$.data"
-    const val ERRORS_JSON_PATH = "$.errors"
-    const val EXTENSIONS_JSON_PATH = "$.extensions"
-    const val GRAPHQL_ENDPOINT = "/graphql"
-    val GRAPHQL_MEDIA_TYPE = MediaType("application", "graphql")
-    const val SUBSCRIPTION_ENDPOINT = "/subscriptions"
-}
+const val DATA_JSON_PATH = "$.data"
+const val ERRORS_JSON_PATH = "$.errors"
+const val EXTENSIONS_JSON_PATH = "$.extensions"
+const val GRAPHQL_ENDPOINT = "/graphql"
+val GRAPHQL_MEDIA_TYPE = MediaType("application", "graphql")
+const val SUBSCRIPTION_ENDPOINT = "/subscriptions"
