+++ examples/spring/src/test/kotlin/com/expediagroup/graphql/examples/mutation/ScalarMutationIT.kt
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package com.expediagroup.graphql.examples.query
+package com.expediagroup.graphql.examples.mutation
 
-import com.expediagroup.graphql.examples.Constants.DATA_JSON_PATH
-import com.expediagroup.graphql.examples.Constants.GRAPHQL_ENDPOINT
-import com.expediagroup.graphql.examples.Constants.GRAPHQL_MEDIA_TYPE
-import com.expediagroup.graphql.examples.IntegrationTest
+import com.expediagroup.graphql.examples.DATA_JSON_PATH
+import com.expediagroup.graphql.examples.GRAPHQL_ENDPOINT
+import com.expediagroup.graphql.examples.GRAPHQL_MEDIA_TYPE
+import com.expediagroup.graphql.examples.verifyOnlyDataExists
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS
@@ -32,21 +32,20 @@ import org.springframework.test.web.reactive.server.WebTestClient
 @SpringBootTest
 @AutoConfigureWebTestClient
 @TestInstance(PER_CLASS)
-class PrivateInterfaceQueryIT(@Autowired private val testClient: WebTestClient) : IntegrationTest {
+class ScalarMutationIT(@Autowired private val testClient: WebTestClient) {
 
     @Test
-    fun `verify queryForObjectWithPrivateInterface query`() {
-        val query = "queryForObjectWithPrivateInterface"
+    fun `verify addPerson query`() {
+        val query = "addPerson"
 
         testClient.post()
             .uri(GRAPHQL_ENDPOINT)
             .accept(APPLICATION_JSON)
             .contentType(GRAPHQL_MEDIA_TYPE)
-            .bodyValue("query { $query { id, value } }")
+            .bodyValue("mutation { $query(person: {id: 1, name: \"Alice\"}) { id, name } }")
             .exchange()
-            .expectStatus().isOk
             .verifyOnlyDataExists(query)
-            .jsonPath("$DATA_JSON_PATH.$query.id").isEqualTo("123")
-            .jsonPath("$DATA_JSON_PATH.$query.value").isEqualTo("Implementation of a method from a private interface")
+            .jsonPath("$DATA_JSON_PATH.$query.id").isEqualTo(1)
+            .jsonPath("$DATA_JSON_PATH.$query.name").isEqualTo("Alice")
     }
 }
