+++ examples/spring/src/test/kotlin/com/expediagroup/graphql/examples/mutation/SimpleMutationIT.kt
@@ -1,0 +1,67 @@
+/*
+ * Copyright 2019 Expedia, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.expediagroup.graphql.examples.mutation
+
+import com.expediagroup.graphql.examples.DATA_JSON_PATH
+import com.expediagroup.graphql.examples.GRAPHQL_ENDPOINT
+import com.expediagroup.graphql.examples.GRAPHQL_MEDIA_TYPE
+import com.expediagroup.graphql.examples.verifyOnlyDataExists
+import org.hamcrest.Matchers
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestInstance
+import org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.http.MediaType.APPLICATION_JSON
+import org.springframework.test.web.reactive.server.WebTestClient
+
+@SpringBootTest
+@AutoConfigureWebTestClient
+@TestInstance(PER_CLASS)
+class SimpleMutationIT(@Autowired private val testClient: WebTestClient) {
+
+    @Test
+    fun `verify addToList query`() {
+        val query = "addToList"
+        val firstEntry = "first entry"
+        val secondEntry = "second entry"
+
+        testClient.post()
+            .uri(GRAPHQL_ENDPOINT)
+            .accept(APPLICATION_JSON)
+            .contentType(GRAPHQL_MEDIA_TYPE)
+            .bodyValue("mutation { $query(entry: \"$firstEntry\") }")
+            .exchange()
+            .verifyOnlyDataExists(query)
+            .jsonPath("$DATA_JSON_PATH.$query").isArray
+            .jsonPath("$DATA_JSON_PATH.$query").value(Matchers.hasSize<Int>(1))
+            .jsonPath("$DATA_JSON_PATH.$query.[0]").isEqualTo(firstEntry)
+
+        testClient.post()
+            .uri(GRAPHQL_ENDPOINT)
+            .accept(APPLICATION_JSON)
+            .contentType(GRAPHQL_MEDIA_TYPE)
+            .bodyValue("mutation { $query(entry: \"$secondEntry\") }")
+            .exchange()
+            .verifyOnlyDataExists(query)
+            .jsonPath("$DATA_JSON_PATH.$query").isArray
+            .jsonPath("$DATA_JSON_PATH.$query").value(Matchers.hasSize<Int>(2))
+            .jsonPath("$DATA_JSON_PATH.$query.[0]").isEqualTo(firstEntry)
+            .jsonPath("$DATA_JSON_PATH.$query.[1]").isEqualTo(secondEntry)
+    }
+}
