+++ examples/spring/src/test/kotlin/com/expediagroup/graphql/examples/query/AsyncQueryIT.kt
@@ -16,9 +16,9 @@
 
 package com.expediagroup.graphql.examples.query
 
-import com.expediagroup.graphql.examples.Constants.GRAPHQL_ENDPOINT
-import com.expediagroup.graphql.examples.Constants.GRAPHQL_MEDIA_TYPE
-import com.expediagroup.graphql.examples.IntegrationTest
+import com.expediagroup.graphql.examples.GRAPHQL_ENDPOINT
+import com.expediagroup.graphql.examples.GRAPHQL_MEDIA_TYPE
+import com.expediagroup.graphql.examples.verifyData
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS
@@ -31,46 +31,18 @@ import org.springframework.test.web.reactive.server.WebTestClient
 @SpringBootTest
 @AutoConfigureWebTestClient
 @TestInstance(PER_CLASS)
-class CoroutineQueryIT(@Autowired private val testClient: WebTestClient) : IntegrationTest {
+class AsyncQueryIT(@Autowired private val testClient: WebTestClient) {
 
     @Test
-    fun `verify exampleCoroutineQuery query`() {
-        val query = "exampleCoroutineQuery"
-        val expectedData = "hello:olleh"
-
-        testClient.post()
-            .uri(GRAPHQL_ENDPOINT)
-            .accept(APPLICATION_JSON)
-            .contentType(GRAPHQL_MEDIA_TYPE)
-            .bodyValue("query { $query(msg: \"hello\") }")
-            .exchange()
-            .verifyData(query, expectedData)
-    }
-
-    @Test
-    fun `verify slowFunction query`() {
-        val query = "slowFunction"
-        val expectedData = "olleh"
-
-        testClient.post()
-            .uri(GRAPHQL_ENDPOINT)
-            .accept(APPLICATION_JSON)
-            .contentType(GRAPHQL_MEDIA_TYPE)
-            .bodyValue("query { $query(msg: \"hello\") }")
-            .exchange()
-            .verifyData(query, expectedData)
-    }
-
-    @Test
-    fun `verify fastFunction query`() {
-        val query = "fastFunction"
+    fun `verify delayedEchoUsingCompletableFuture query`() {
+        val query = "delayedEchoUsingCompletableFuture"
         val data = "hello"
 
         testClient.post()
             .uri(GRAPHQL_ENDPOINT)
             .accept(APPLICATION_JSON)
             .contentType(GRAPHQL_MEDIA_TYPE)
-            .bodyValue("query { $query(msg: \"$data\") }")
+            .bodyValue("query { $query(msg: \"$data\", delaySeconds: 1) }")
             .exchange()
             .verifyData(query, data)
     }
