+++ examples/spring/src/test/kotlin/com/expediagroup/graphql/examples/query/NestedQueriesIT.kt
@@ -16,10 +16,10 @@
 
 package com.expediagroup.graphql.examples.query
 
-import com.expediagroup.graphql.examples.Constants.DATA_JSON_PATH
-import com.expediagroup.graphql.examples.Constants.GRAPHQL_ENDPOINT
-import com.expediagroup.graphql.examples.Constants.GRAPHQL_MEDIA_TYPE
-import com.expediagroup.graphql.examples.IntegrationTest
+import com.expediagroup.graphql.examples.DATA_JSON_PATH
+import com.expediagroup.graphql.examples.GRAPHQL_ENDPOINT
+import com.expediagroup.graphql.examples.GRAPHQL_MEDIA_TYPE
+import com.expediagroup.graphql.examples.verifyOnlyDataExists
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS
@@ -32,38 +32,39 @@ import org.springframework.test.web.reactive.server.WebTestClient
 @SpringBootTest
 @AutoConfigureWebTestClient
 @TestInstance(PER_CLASS)
-class ContextualQueryIT(@Autowired private val testClient: WebTestClient) : IntegrationTest {
+class NestedQueriesIT(@Autowired private val testClient: WebTestClient) {
 
     @Test
-    fun `verify contextualQuery query`() {
-        val query = "contextualQuery"
+    fun `verify findAnimal query`() {
+        val query = "findAnimal"
 
         testClient.post()
             .uri(GRAPHQL_ENDPOINT)
             .accept(APPLICATION_JSON)
             .contentType(GRAPHQL_MEDIA_TYPE)
-            .bodyValue("query { $query(value: 1) { contextValue, passedInValue } }")
+            .bodyValue("query { $query { details { veryDetailedFunction }, id, type } }")
             .exchange()
             .expectStatus().isOk
             .verifyOnlyDataExists(query)
-            .jsonPath("$DATA_JSON_PATH.$query.contextValue").isEqualTo("defaultContext")
-            .jsonPath("$DATA_JSON_PATH.$query.passedInValue").isEqualTo("1")
+            .jsonPath("$DATA_JSON_PATH.$query.details").exists()
+            .jsonPath("$DATA_JSON_PATH.$query.details.veryDetailedFunction").isEqualTo("Details(1)")
+            .jsonPath("$DATA_JSON_PATH.$query.id").isEqualTo("1")
+            .jsonPath("$DATA_JSON_PATH.$query.type").isEqualTo("cat")
     }
 
     @Test
-    fun `verify contextualQuery query with context value`() {
-        val query = "contextualQuery"
+    fun `verify getCoffeeBean query`() {
+        val query = "getCoffeeBean"
 
         testClient.post()
             .uri(GRAPHQL_ENDPOINT)
-            .header("MyHeader", "hello")
             .accept(APPLICATION_JSON)
             .contentType(GRAPHQL_MEDIA_TYPE)
-            .bodyValue("query { $query(value: 1) { contextValue, passedInValue } }")
+            .bodyValue("query { $query(beanName: \"arabica\") { hotCoffee(size: \"small\"), icedCoffee(size: \"large\") } }")
             .exchange()
             .expectStatus().isOk
             .verifyOnlyDataExists(query)
-            .jsonPath("$DATA_JSON_PATH.$query.contextValue").isEqualTo("hello")
-            .jsonPath("$DATA_JSON_PATH.$query.passedInValue").isEqualTo("1")
+            .jsonPath("$DATA_JSON_PATH.$query.hotCoffee").isEqualTo("Hot Coffee, Bean choice: arabica, size: small")
+            .jsonPath("$DATA_JSON_PATH.$query.icedCoffee").isEqualTo("Iced Coffee, Bean choice: arabica, size: large")
     }
 }
