+++ examples/spring/src/test/kotlin/com/expediagroup/graphql/examples/query/PrivateInterfaceQueryIT.kt
@@ -16,10 +16,10 @@
 
 package com.expediagroup.graphql.examples.query
 
-import com.expediagroup.graphql.examples.Constants.DATA_JSON_PATH
-import com.expediagroup.graphql.examples.Constants.GRAPHQL_ENDPOINT
-import com.expediagroup.graphql.examples.Constants.GRAPHQL_MEDIA_TYPE
-import com.expediagroup.graphql.examples.IntegrationTest
+import com.expediagroup.graphql.examples.DATA_JSON_PATH
+import com.expediagroup.graphql.examples.GRAPHQL_ENDPOINT
+import com.expediagroup.graphql.examples.GRAPHQL_MEDIA_TYPE
+import com.expediagroup.graphql.examples.verifyOnlyDataExists
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS
@@ -32,19 +32,21 @@ import org.springframework.test.web.reactive.server.WebTestClient
 @SpringBootTest
 @AutoConfigureWebTestClient
 @TestInstance(PER_CLASS)
-class SubQueriesIT(@Autowired private val testClient: WebTestClient) : IntegrationTest {
+class PrivateInterfaceQueryIT(@Autowired private val testClient: WebTestClient) {
 
     @Test
-    fun `verify main query`() {
-        val query = "main"
+    fun `verify queryForObjectWithPrivateInterface query`() {
+        val query = "queryForObjectWithPrivateInterface"
 
         testClient.post()
             .uri(GRAPHQL_ENDPOINT)
             .accept(APPLICATION_JSON)
             .contentType(GRAPHQL_MEDIA_TYPE)
-            .bodyValue("query { $query { secondary } }")
+            .bodyValue("query { $query { id, value } }")
             .exchange()
+            .expectStatus().isOk
             .verifyOnlyDataExists(query)
-            .jsonPath("$DATA_JSON_PATH.$query.secondary").isEqualTo("secondary")
+            .jsonPath("$DATA_JSON_PATH.$query.id").isEqualTo("123")
+            .jsonPath("$DATA_JSON_PATH.$query.value").isEqualTo("Implementation of a method from a private interface")
     }
 }
