+++ examples/spring/src/test/kotlin/com/expediagroup/graphql/examples/query/ScalarQueryIT.kt
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package com.expediagroup.graphql.examples.mutation
+package com.expediagroup.graphql.examples.query
 
-import com.expediagroup.graphql.examples.Constants.DATA_JSON_PATH
-import com.expediagroup.graphql.examples.Constants.GRAPHQL_ENDPOINT
-import com.expediagroup.graphql.examples.Constants.GRAPHQL_MEDIA_TYPE
-import com.expediagroup.graphql.examples.IntegrationTest
-import org.hamcrest.Matchers
+import com.expediagroup.graphql.examples.DATA_JSON_PATH
+import com.expediagroup.graphql.examples.GRAPHQL_ENDPOINT
+import com.expediagroup.graphql.examples.GRAPHQL_MEDIA_TYPE
+import com.expediagroup.graphql.examples.verifyOnlyDataExists
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS
@@ -33,35 +32,34 @@ import org.springframework.test.web.reactive.server.WebTestClient
 @SpringBootTest
 @AutoConfigureWebTestClient
 @TestInstance(PER_CLASS)
-class SimpleMutationIT(@Autowired private val testClient: WebTestClient) : IntegrationTest {
+class ScalarQueryIT(@Autowired private val testClient: WebTestClient) {
 
     @Test
-    fun `verify addToList query`() {
-        val query = "addToList"
-        val firstEntry = "first entry"
-        val secondEntry = "second entry"
+    fun `verify generateRandomUUID query`() {
+        val query = "generateRandomUUID"
 
         testClient.post()
             .uri(GRAPHQL_ENDPOINT)
             .accept(APPLICATION_JSON)
             .contentType(GRAPHQL_MEDIA_TYPE)
-            .bodyValue("mutation { $query(entry: \"$firstEntry\") }")
+            .bodyValue("query { $query }")
             .exchange()
             .verifyOnlyDataExists(query)
-            .jsonPath("$DATA_JSON_PATH.$query").isArray
-            .jsonPath("$DATA_JSON_PATH.$query").value(Matchers.hasSize<Int>(1))
-            .jsonPath("$DATA_JSON_PATH.$query.[0]").isEqualTo(firstEntry)
+            .jsonPath("$DATA_JSON_PATH.$query").isNotEmpty
+    }
+
+    @Test
+    fun `verify findPersonById query`() {
+        val query = "findPersonById"
 
         testClient.post()
             .uri(GRAPHQL_ENDPOINT)
             .accept(APPLICATION_JSON)
             .contentType(GRAPHQL_MEDIA_TYPE)
-            .bodyValue("mutation { $query(entry: \"$secondEntry\") }")
+            .bodyValue("query { $query(id: \"1\") { id, name } }")
             .exchange()
             .verifyOnlyDataExists(query)
-            .jsonPath("$DATA_JSON_PATH.$query").isArray
-            .jsonPath("$DATA_JSON_PATH.$query").value(Matchers.hasSize<Int>(2))
-            .jsonPath("$DATA_JSON_PATH.$query.[0]").isEqualTo(firstEntry)
-            .jsonPath("$DATA_JSON_PATH.$query.[1]").isEqualTo(secondEntry)
+            .jsonPath("$DATA_JSON_PATH.$query.id").isEqualTo("1")
+            .jsonPath("$DATA_JSON_PATH.$query.name").isEqualTo("Nelson")
     }
 }
