+++ examples/spring/src/test/kotlin/com/expediagroup/graphql/examples/query/ValidatedQueryIT.kt
@@ -16,12 +16,15 @@
 
 package com.expediagroup.graphql.examples.query
 
-import com.expediagroup.graphql.examples.Constants.GRAPHQL_ENDPOINT
-import com.expediagroup.graphql.examples.Constants.GRAPHQL_MEDIA_TYPE
-import com.expediagroup.graphql.examples.IntegrationTest
+import com.expediagroup.graphql.examples.GRAPHQL_ENDPOINT
+import com.expediagroup.graphql.examples.GRAPHQL_MEDIA_TYPE
+import com.expediagroup.graphql.examples.verifyData
+import com.expediagroup.graphql.examples.verifyError
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.ValueSource
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient
 import org.springframework.boot.test.context.SpringBootTest
@@ -31,19 +34,35 @@ import org.springframework.test.web.reactive.server.WebTestClient
 @SpringBootTest
 @AutoConfigureWebTestClient
 @TestInstance(PER_CLASS)
-class AsyncQueryIT(@Autowired private val testClient: WebTestClient) : IntegrationTest {
+class ValidatedQueryIT(@Autowired private val testClient: WebTestClient) {
 
     @Test
-    fun `verify delayedEchoUsingCompletableFuture query`() {
-        val query = "delayedEchoUsingCompletableFuture"
-        val data = "hello"
+    fun `verify argumentWithValidation query`() {
+        val query = "argumentWithValidation"
+        val argument = "hello"
 
         testClient.post()
             .uri(GRAPHQL_ENDPOINT)
             .accept(APPLICATION_JSON)
             .contentType(GRAPHQL_MEDIA_TYPE)
-            .bodyValue("query { $query(msg: \"$data\", delaySeconds: 1) }")
+            .bodyValue("query { $query(arg: { lowerCaseOnly: \"$argument\" }) }")
             .exchange()
-            .verifyData(query, data)
+            .verifyData(query, argument)
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = ["", " ", "Hello", "1234", "!@#$"])
+    fun `verify argumentWithValidation query when argument is not valid`(argument: String) {
+        val query = "argumentWithValidation"
+        val expectedError = "Exception while fetching data (argumentWithValidation) : " +
+            "argumentWithValidation.arg.lowerCaseOnly: Argument must be lowercase"
+
+        testClient.post()
+            .uri(GRAPHQL_ENDPOINT)
+            .accept(APPLICATION_JSON)
+            .contentType(GRAPHQL_MEDIA_TYPE)
+            .bodyValue("query { $query(arg: { lowerCaseOnly: \"$argument\" }) }")
+            .exchange()
+            .verifyError(expectedError)
     }
 }
